FormatParams::parse_line
